backend/requirements.txt
# MarkezardAI Production Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.2.0
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
bleach==6.1.0
aiohttp==3.9.1
python-dotenv==1.0.0
httpx==0.25.2
